requirements.txt
flask==3.0.0
transformers>=4.40.0
torch>=2.0.0
tweepy>=4.15.0
praw==7.7.1
slack-sdk==3.26.1
python-dotenv==1.0.0
APScheduler==3.10.4
requests==2.31.0
sentencepiece

